requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
pandas>=2.0.0
jobspy>=0.0.1
google-generativeai>=0.3.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0 